chore: 프로젝트명 savit으로 변경 (settings.gradle) (#17)
