--- src/main/java/com/oroom/config/ServletConfig.java
@@ -1,12 +1,0 @@
-package com.oroom.config;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-
-@Configuration
-@EnableWebMvc
-@ComponentScan(basePackages = {
-        "com.oroom.user.controller"  // 컨트롤러들만
-})public class ServletConfig {
-}
