--- src/main/java/com/oroom/security/TokenPairDTO.java
@@ -1,20 +1,0 @@
-package com.oroom.security;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class TokenPairDTO {
-
-    private String accessToken;
-    private String refreshToken;
-
-    // 토큰 만료 시간 정보
-    private Long accessTokenExpiresIn;
-    private Long refreshTokenExpiresIn;
-}
