--- src/main/java/com/oroom/user/domain/User.java
@@ -1,30 +1,0 @@
-package com.oroom.user.domain;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class User {
-    private Long id;
-    private String email;
-    private String nickname;
-    private String profileImage;
-
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    private LocalDateTime createdAt;
-
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    private LocalDateTime updatedAt;
-    private String kakaoUserId;
-    private String connectedId;
-    private String refreshToken;
-    private String birthDate;
-}
