--- src/main/java/com/oroom/user/dto/KakaoTokenDTO.java
@@ -1,16 +1,0 @@
-package com.oroom.user.dto;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategy;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.Data;
-
-@Data
-@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
-public class KakaoTokenDTO {
-    private String accessToken;
-    private String tokenType;
-    private String refreshToken;
-    private int expiresIn;
-    private String scope;
-    private int refreshTokenExpiresIn;
-}
