--- src/main/java/com/oroom/user/dto/UserDTO.java
@@ -1,19 +1,0 @@
-package com.oroom.user.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-public class UserDTO {
-    private Long id;
-    private String email;
-    private String nickname;
-    private String profileImage;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    private String kakaoUserId;
-
-}
