--- src/main/java/com/oroom/user/service/KakaoOAuthService.java
@@ -1,9 +1,0 @@
-package com.oroom.user.service;
-
-import com.oroom.user.dto.KakaoTokenDTO;
-import com.oroom.user.dto.KakaoUserDTO;
-
-public interface KakaoOAuthService {
-    KakaoTokenDTO getAccessToken(String code);
-    KakaoUserDTO getUserInfo(String accessToken);
-}
