--- src/main/java/com/oroom/user/service/UserService.java
@@ -1,10 +1,0 @@
-package com.oroom.user.service;
-
-import com.oroom.user.domain.User;
-
-public interface UserService {
-    User findByEmail(String email);
-    User findByKakaoUserId(String kakaoUserId);
-    void createUser(User user);
-    void updateUser(User user);
-}
