--- src/main/java/com/oroom/user/service/UserServiceImpl.java
@@ -1,36 +1,0 @@
-package com.oroom.user.service;
-
-import com.oroom.user.domain.User;
-import com.oroom.user.mapper.UserMapper;
-import com.oroom.user.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional
-public class UserServiceImpl implements UserService {
-
-    private final UserMapper userMapper;
-
-    @Override
-    public User findByEmail(String email) {
-        return userMapper.findByEmail(email);
-    }
-
-    @Override
-    public User findByKakaoUserId(String kakaoUserId) {
-        return userMapper.findByKakaoUserId(kakaoUserId);
-    }
-
-    @Override
-    public void createUser(User user) {
-        userMapper.insertUser(user);
-    }
-
-    @Override
-    public void updateUser(User user) {
-        userMapper.updateUser(user);
-    }
-}
