+++ src/main/java/com/savit/config/RootConfig.java
@@ -1,0 +1,88 @@
+package com.savit.config;
+
+import com.zaxxer.hikari.HikariConfig;
+import com.zaxxer.hikari.HikariDataSource;
+import org.apache.ibatis.session.SqlSessionFactory;
+import org.mybatis.spring.SqlSessionFactoryBean;
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
+import org.springframework.jdbc.datasource.DataSourceTransactionManager;
+
+import javax.sql.DataSource;
+
+@Configuration
+@PropertySource("classpath:/application.properties")
+@MapperScan("com.savit.user.mapper")  // 매퍼만 스캔
+@ComponentScan(basePackages = {
+        "com.savit.user.service",    // 서비스 클래스들
+        "com.savit.security"         // KakaoProperties, JwtTokenProvider
+})
+public class RootConfig {
+
+    @Bean
+    public static PropertySourcesPlaceholderConfigurer propertyConfig() {
+        return new PropertySourcesPlaceholderConfigurer();
+    }
+
+    @Value("${jdbc.driver}")
+    private String driver;
+
+    @Value("${jdbc.url}")
+    private String url;
+
+    @Value("${jdbc.username}")
+    private String username;
+
+    @Value("${jdbc.password}")
+    private String password;
+
+    private final ApplicationContext applicationContext;
+
+    public RootConfig(ApplicationContext applicationContext) {
+        this.applicationContext = applicationContext;
+    }
+
+    // DataSource 설정 (MySQL + HikariCP)
+    @Bean
+    public DataSource dataSource() {
+        HikariConfig config = new HikariConfig();
+        config.setDriverClassName(driver);
+        config.setJdbcUrl(url);
+        config.setUsername(username);
+        config.setPassword(password);
+        config.setMaximumPoolSize(10);
+        config.setMinimumIdle(2);
+        config.setPoolName("OroomHikariPool");
+        return new HikariDataSource(config);
+    }
+
+    // SqlSessionFactory 설정
+    @Bean
+    public SqlSessionFactory sqlSessionFactory() throws Exception {
+        SqlSessionFactoryBean factory = new SqlSessionFactoryBean();
+        factory.setDataSource(dataSource());
+        factory.setConfigLocation(applicationContext.getResource("classpath:/mybatis-config.xml"));
+
+        // 매퍼 위치 직접 지정 - 이 부분을 추가!
+        factory.setMapperLocations(
+                new PathMatchingResourcePatternResolver().getResources("classpath:mapper/**/*.xml")
+        );
+
+        // 타입 별칭 패키지 설정
+        factory.setTypeAliasesPackage("com.oroom.user.domain");
+
+        return factory.getObject();
+    }
+    // 트랜잭션 매니저 설정
+    @Bean
+    public DataSourceTransactionManager transactionManager() {
+        return new DataSourceTransactionManager(dataSource());
+    }
+}
