+++ src/main/java/com/savit/config/ServletConfig.java
@@ -1,0 +1,12 @@
+package com.savit.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+
+@Configuration
+@EnableWebMvc
+@ComponentScan(basePackages = {
+        "com.savit.user.controller"  // 컨트롤러들만
+})public class ServletConfig {
+}
