+++ src/main/java/com/savit/config/SwaggerConfig.java
@@ -1,0 +1,41 @@
+package com.savit.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.bind.annotation.RestController;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@Configuration
+@Profile("!test")
+@EnableSwagger2
+public class SwaggerConfig {
+
+    private final String API_NAME = "ORoom";
+    private final String API_VERSION = "1.0";
+    private final String API_DESCRIPTION = "KB 6기 오룸 프로젝트";
+
+    private ApiInfo apiInfo() {
+        return new ApiInfoBuilder()
+                .title(API_NAME)
+                .description(API_DESCRIPTION)
+                .version(API_VERSION)
+                .build();
+    }
+
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.withClassAnnotation(RestController.class))
+                .paths(PathSelectors.any())
+                .build()
+                .apiInfo(apiInfo());
+    }
+}
