+++ src/main/java/com/savit/security/KakaoProperties.java
@@ -1,0 +1,23 @@
+package com.savit.security;
+
+import lombok.Getter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+
+@Component
+@Getter
+public class KakaoProperties {
+
+    @Value("${kakao.client-id}")
+    private String clientId;
+
+    @Value("${kakao.redirect-uri}")
+    private String redirectUri;
+
+    @Value("${kakao.token-url}")
+    private String tokenUrl;
+
+    @Value("${kakao.user-info-url}")
+    private String userInfoUrl;
+}
