+++ src/main/java/com/savit/user/domain/User.java
@@ -1,0 +1,30 @@
+package com.savit.user.domain;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class User {
+    private Long id;
+    private String email;
+    private String nickname;
+    private String profileImage;
+
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime createdAt;
+
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime updatedAt;
+    private String kakaoUserId;
+    private String connectedId;
+    private String refreshToken;
+    private String birthDate;
+}
