+++ src/main/java/com/savit/user/dto/KakaoTokenDTO.java
@@ -1,0 +1,16 @@
+package com.savit.user.dto;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Data;
+
+@Data
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+public class KakaoTokenDTO {
+    private String accessToken;
+    private String tokenType;
+    private String refreshToken;
+    private int expiresIn;
+    private String scope;
+    private int refreshTokenExpiresIn;
+}
