+++ src/main/java/com/savit/user/dto/KakaoUserDTO.java
@@ -1,0 +1,36 @@
+package com.savit.user.dto;
+
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Data;
+
+@Data
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+public class KakaoUserDTO {
+    private Long id;
+    private KakaoAccount kakaoAccount;
+    private Properties properties;
+
+    @Data
+    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+    public static class KakaoAccount {
+        private String email;
+        private Profile profile;
+
+        @Data
+        @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+        public static class Profile {
+            private String nickname;
+            private String profileImageUrl;
+        }
+    }
+
+    @Data
+    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+    public static class Properties {
+        private String nickname;
+        private String profileImage;
+    }
+}
