+++ src/main/java/com/savit/user/service/KakaoOAuthService.java
@@ -1,0 +1,9 @@
+package com.savit.user.service;
+
+import com.savit.user.dto.KakaoTokenDTO;
+import com.savit.user.dto.KakaoUserDTO;
+
+public interface KakaoOAuthService {
+    KakaoTokenDTO getAccessToken(String code);
+    KakaoUserDTO getUserInfo(String accessToken);
+}
