+++ src/main/java/com/savit/user/service/UserService.java
@@ -1,0 +1,10 @@
+package com.savit.user.service;
+
+import com.savit.user.domain.User;
+
+public interface UserService {
+    User findByEmail(String email);
+    User findByKakaoUserId(String kakaoUserId);
+    void createUser(User user);
+    void updateUser(User user);
+}
